requirement.txt
dashscope==1.15.0
langchain-core==0.3.12
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.3
matplotlib==3.8.3
pillow==10.2.0
numpy==1.26.4
openai==1.52.0
pandas==1.5.3
sqlalchemy==2.0.29
pyyaml==6.0.1
uvicorn~=0.31.0
fastapi~=0.115.0
pydantic~=2.9.2
pyecharts==2.0.6
pygwalker==0.4.8.7
PyMySQL==1.1.1
sentence-transformers==3.0.1
pgvector==0.2.5
psycopg2-binary==2.9.9